requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
dnspython==2.3.0
docx==0.2.4
environ==1.0
filetype==1.2.0
idna==3.4
lxml==4.9.2
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
pymongo==4.3.3
PyMuPDF==1.22.5
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
requests==2.30.0
six==1.16.0
termcolor==2.3.0
tzdata==2023.3
urllib3==2.0.2
django-environ
python-magic==0.4.27

